Final Refactoring. Renaming and removing line spaces
